Extract `WalletAddress` + `WalletType` to `shared`
