+++ src/main/api/ledger/binance/address.ts
@@ -5,10 +5,10 @@ import { getDerivePath, getPrefix } from '@xchainjs/xchain-binance'
 import { BNBChain } from '@xchainjs/xchain-util'
 import * as E from 'fp-ts/Either'
 
-import { WalletAddress } from '../../../../renderer/services/wallet/types'
 import { LedgerError, LedgerErrorId, Network } from '../../../../shared/api/types'
 import { toClientNetwork } from '../../../../shared/utils/client'
 import { isError } from '../../../../shared/utils/guard'
+import { WalletAddress } from '../../../../shared/wallet/types'
 
 export const getAddress = async (
   transport: Transport,
