+++ src/renderer/components/AccountAddressSelector/AccountAddressSelector.types.ts
@@ -1,7 +1,7 @@
 import { Address } from '@xchainjs/xchain-client'
 import { Chain } from '@xchainjs/xchain-util'
 
-import { WalletType } from '../../services/wallet/types'
+import { WalletType } from '../../../shared/wallet/types'
 
 export type AccountAddressSelectorType = {
   walletAddress: Address
