+++ src/renderer/components/uielements/assets/assetInfo/AssetInfo.tsx
@@ -9,10 +9,11 @@ import { useIntl } from 'react-intl'
 
 import { Network } from '../../../../../shared/api/types'
 import { isLedgerWallet } from '../../../../../shared/utils/guard'
+import { WalletType } from '../../../../../shared/wallet/types'
 import { sequenceSOption, sequenceTOption } from '../../../../helpers/fpHelpers'
 import { loadingString, emptyString } from '../../../../helpers/stringHelper'
 import { getAssetAmountByAsset } from '../../../../helpers/walletHelper'
-import { NonEmptyWalletBalances, WalletType } from '../../../../services/wallet/types'
+import { NonEmptyWalletBalances } from '../../../../services/wallet/types'
 import { QRCodeModal } from '../../qrCodeModal/QRCodeModal'
 import { AssetIcon } from '../assetIcon'
 import * as Styled from './AssetInfo.styles'
