+++ src/renderer/components/wallet/account/AccountSelector.tsx
@@ -7,8 +7,9 @@ import { useIntl } from 'react-intl'
 
 import { Network } from '../../../../shared/api/types'
 import { isLedgerWallet } from '../../../../shared/utils/guard'
+import { WalletType } from '../../../../shared/wallet/types'
 import { WalletBalances } from '../../../services/clients'
-import { WalletBalance, WalletType } from '../../../services/wallet/types'
+import { WalletBalance } from '../../../services/wallet/types'
 import { AssetData } from '../../uielements/assets/assetData'
 import { AssetIcon } from '../../uielements/assets/assetIcon'
 import { Size as IconSize } from '../../uielements/assets/assetIcon/AssetIcon.types'
