+++ src/renderer/components/wallet/txs/send/Send.stories.tsx
@@ -22,9 +22,10 @@ import { Network } from '../../../../../shared/api/types'
 import { BNB_TRANSFER_FEES } from '../../../../../shared/mock/fees'
 import { RDStatus, getMockRDValueFactory } from '../../../../../shared/mock/rdByStatus'
 import { mockValidatePassword$ } from '../../../../../shared/mock/wallet'
+import { WalletType } from '../../../../../shared/wallet/types'
 import { SendTxParams } from '../../../../services/binance/types'
 import { WalletBalances } from '../../../../services/clients'
-import { ErrorId, TxHashRD, WalletBalance, WalletType } from '../../../../services/wallet/types'
+import { ErrorId, TxHashRD, WalletBalance } from '../../../../services/wallet/types'
 import { Send } from './Send'
 import { SendFormBNB } from './SendFormBNB'
 import { SendFormBTC } from './SendFormBTC'
