+++ src/renderer/helpers/walletHelper.ts
@@ -4,9 +4,10 @@ import * as A from 'fp-ts/Array'
 import * as FP from 'fp-ts/function'
 import * as O from 'fp-ts/Option'
 
+import { WalletAddress } from '../../shared/wallet/types'
 import { ZERO_ASSET_AMOUNT } from '../const'
 import { WalletBalances } from '../services/clients'
-import { NonEmptyWalletBalances, WalletAddress, WalletBalance } from '../services/wallet/types'
+import { NonEmptyWalletBalances, WalletBalance } from '../services/wallet/types'
 import { isBnbAsset, isEthAsset, isLtcAsset, isRuneNativeAsset } from './assetHelper'
 import { eqAddress, eqAsset } from './fp/eq'
 import { sequenceTOption } from './fpHelpers'
