+++ src/renderer/routes/wallet/wallet.ts
@@ -4,9 +4,9 @@ import * as FP from 'fp-ts/lib/function'
 import * as O from 'fp-ts/lib/Option'
 
 import { Network } from '../../../shared/api/types'
+import { WalletType } from '../../../shared/wallet/types'
 import { isNonNativeRuneAsset } from '../../helpers/assetHelper'
 import { sequenceTOption } from '../../helpers/fpHelpers'
-import { WalletType } from '../../services/wallet/types'
 import { Route } from '../types'
 
 export * as imports from './imports'
