+++ src/renderer/services/thorchain/balances.ts
@@ -1,8 +1,8 @@
 import { AssetRuneNative } from '@xchainjs/xchain-util'
 
+import { WalletType } from '../../../shared/wallet/types'
 import { observableState } from '../../helpers/stateHelper'
 import * as C from '../clients'
-import { WalletType } from '../wallet/types'
 import { client$ } from './common'
 
 /**
