+++ src/renderer/services/thorchain/interact.ts
@@ -7,10 +7,11 @@ import * as O from 'fp-ts/lib/Option'
 import * as Rx from 'rxjs'
 import * as RxOp from 'rxjs/operators'
 
+import { WalletType } from '../../../shared/wallet/types'
 import { liveData } from '../../helpers/rx/liveData'
 import { LiveData } from '../../helpers/rx/liveData'
 import { observableState } from '../../helpers/stateHelper'
-import { TxLD, WalletType } from '../wallet/types'
+import { TxLD } from '../wallet/types'
 import { ApiError } from '../wallet/types'
 import { INITIAL_INTERACT_STATE } from './const'
 import { InteractParams, InteractState, InteractState$ } from './types'
