+++ src/renderer/services/wallet/types.ts
@@ -9,6 +9,7 @@ import * as O from 'fp-ts/lib/Option'
 import * as Rx from 'rxjs'
 
 import { LedgerError, Network } from '../../../shared/api/types'
+import { WalletAddress, WalletType } from '../../../shared/wallet/types'
 import { LiveData } from '../../helpers/rx/liveData'
 import { LoadTxsParams, WalletBalancesLD, WalletBalancesRD } from '../clients'
 
@@ -46,10 +47,6 @@ export type KeystoreService = {
    */
   validatePassword$: ValidatePasswordHandler
 }
-
-export type WalletType = 'keystore' | 'ledger'
-
-export type WalletAddress = { address: Address; type: WalletType; chain: Chain }
 
 export type WalletAddressAsync = { address: RD.RemoteData<Error, WalletAddress>; type: WalletType }
 export type WalletAddressesAsync = WalletAddressAsync[]
