+++ src/renderer/views/wallet/Interact/CustomView.tsx
@@ -7,13 +7,13 @@ import * as FP from 'fp-ts/function'
 import * as O from 'fp-ts/Option'
 import { useIntl } from 'react-intl'
 
+import { WalletType } from '../../../../shared/wallet/types'
 import { Custom } from '../../../components/interact/forms/Custom'
 import { Button } from '../../../components/uielements/button'
 import { useThorchainContext } from '../../../contexts/ThorchainContext'
 import { useSubscriptionState } from '../../../hooks/useSubscriptionState'
 import { INITIAL_INTERACT_STATE } from '../../../services/thorchain/const'
 import { InteractState } from '../../../services/thorchain/types'
-import { WalletType } from '../../../services/wallet/types'
 import * as Styled from './InteractView.styles'
 
 type Props = {
