+++ src/renderer/views/wallet/WalletSettingsView.tsx
@@ -13,6 +13,7 @@ import * as Rx from 'rxjs'
 import * as RxOp from 'rxjs/operators'
 
 import { Network } from '../../../shared/api/types'
+import { WalletAddress } from '../../../shared/wallet/types'
 import { WalletSettings } from '../../components/wallet/settings/'
 import { useAppContext } from '../../contexts/AppContext'
 import { useBinanceContext } from '../../contexts/BinanceContext'
@@ -27,7 +28,7 @@ import { filterEnabledChains, isBnbChain, isThorChain } from '../../helpers/chai
 import { sequenceTOptionFromArray } from '../../helpers/fpHelpers'
 import { useLedger } from '../../hooks/useLedger'
 import { DEFAULT_NETWORK } from '../../services/const'
-import { WalletAddress, WalletAddressAsync } from '../../services/wallet/types'
+import { WalletAddressAsync } from '../../services/wallet/types'
 import { ledgerErrorIdToI18n } from '../../services/wallet/util'
 import { getPhrase } from '../../services/wallet/util'
 import { walletAccount$ } from './WalletSettingsView.helper'
