+++ src/renderer/views/wallet/send/SendViewLTC.tsx
@@ -9,6 +9,7 @@ import { useIntl } from 'react-intl'
 import { useHistory } from 'react-router-dom'
 
 import { Network } from '../../../../shared/api/types'
+import { WalletType } from '../../../../shared/wallet/types'
 import { Send } from '../../../components/wallet/txs/send/'
 import { SendFormLTC } from '../../../components/wallet/txs/send/'
 import { useChainContext } from '../../../contexts/ChainContext'
@@ -20,12 +21,7 @@ import { INITIAL_SEND_STATE } from '../../../services/chain/const'
 import { SendTxParams, SendTxState } from '../../../services/chain/types'
 import { OpenExplorerTxUrl, WalletBalances } from '../../../services/clients'
 import { FeesWithRatesLD } from '../../../services/litecoin/types'
-import {
-  NonEmptyWalletBalances,
-  ValidatePasswordHandler,
-  WalletBalance,
-  WalletType
-} from '../../../services/wallet/types'
+import { NonEmptyWalletBalances, ValidatePasswordHandler, WalletBalance } from '../../../services/wallet/types'
 import * as Helper from './SendView.helper'
 
 type Props = {
