+++ src/renderer/views/wallet/upgrade/types.ts
@@ -3,10 +3,11 @@ import * as FP from 'fp-ts/function'
 import * as O from 'fp-ts/Option'
 
 import { Network } from '../../../../shared/api/types'
+import { WalletType } from '../../../../shared/wallet/types'
 import { UpgradeRuneParams, UpgradeRuneTxState$ } from '../../../services/chain/types'
 import { AddressValidation, OpenExplorerTxUrl } from '../../../services/clients'
 import { PoolAddressRD } from '../../../services/midgard/types'
-import { NonEmptyWalletBalances, ValidatePasswordHandler, WalletType } from '../../../services/wallet/types'
+import { NonEmptyWalletBalances, ValidatePasswordHandler } from '../../../services/wallet/types'
 import { AssetWithDecimal } from '../../../types/asgardex'
 
 export type CommonUpgradeProps = {
