+++ src/shared/wallet/types.ts
@@ -1,0 +1,6 @@
+import { Address } from '@xchainjs/xchain-client'
+import { Chain } from '@xchainjs/xchain-util'
+
+export type WalletType = 'keystore' | 'ledger'
+
+export type WalletAddress = { address: Address; type: WalletType; chain: Chain }
